feat: настроена работа со списками желаний

+++ src/wishes/wishes.service.ts
@@ -1,6 +1,11 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ConflictException,
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateWishDto } from './dto/create-wish.dto';
-import { Repository } from 'typeorm';
+import { In, Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UsersService } from 'src/users/users.service';
 import { Wish } from './entities/wish.entity';
@@ -12,8 +17,8 @@ export class WishesService {
     private usersService: UsersService,
   ) {}
 
-  async createWish(createWishDto: CreateWishDto, userData: any) {
-    const user = await this.usersService.findOneByUsername(userData.username);
+  async createWish(createWishDto: CreateWishDto, username: string) {
+    const user = await this.usersService.getUserByUsername(username);
     const wish = this.wishRepo.create(createWishDto);
     wish.owner = user;
     return this.wishRepo.save(wish);
@@ -27,23 +32,54 @@ export class WishesService {
     return this.wishRepo.find({ order: { copied: 'desc' }, take: 10 });
   }
 
-  async getWishById(id: number) {
-    // const wish = await this.wishRepo.findOneBy({ id });
-    const wish = this.wishRepo.findOne({ where: { id }, relations: ['owner'] });
+  async getWishById(id: number, relations = null) {
+    const wish = this.wishRepo.findOne({ where: { id }, relations });
     if (!wish) {
       throw new NotFoundException();
-      v;
     }
     return wish;
   }
 
-  async deleteWishById(id: number) {
+  async deleteWishById(id: number, currentUser) {
     const wish = await this.wishRepo.findOneBy({ id });
-    console.log(wish);
     if (!wish) {
       throw new NotFoundException();
     }
-
+    if (wish.owner !== currentUser.id) {
+      throw new ForbiddenException();
+    }
     return this.wishRepo.delete({ id });
+  }
+
+  async findWishesById(idArr: any) {
+    const wishes = await this.wishRepo.find({
+      where: { id: In(idArr) },
+    });
+    return wishes;
+  }
+
+  async copyWish(id: number, currentUser) {
+    const wish = await this.getWishById(id, ['owner']);
+    const wishCopy = {
+      name: wish.name,
+      image: wish.image,
+      link: wish.link,
+      price: wish.price,
+      description: wish.description,
+    };
+    const hasWish = await this.wishRepo.find({
+      where: {
+        owner: { _id: currentUser.id },
+      },
+    });
+
+    if (hasWish) {
+      throw new ConflictException();
+    }
+
+    await this.createWish(wishCopy, currentUser.username);
+    await this.wishRepo.update(id, {
+      copied: wish.copied + 1,
+    });
   }
 }
